refactor(server): extract UTM helpers into src/server/utm.js
Pure move, no behavior change. resolveUtmParams + buildUtmString (both pure,
zero external deps) move to src/server/utm.js; server.js imports them. All
three call sites unchanged.

Adds test/utm.test.js (placeholder substitution incl. the channel->key
fallback, and url-encoding) — first real unit coverage, and it verifies the
module loads cleanly in CI (the route guard + node --check don't boot the app,
so a unit test is how we catch a broken extraction here).

Route inventory unchanged (213). node --check clean on server.js + utm.js.
